requirements.txt
streamlit==1.37.0
openai==1.51.0
langchain==0.0.316
tiktoken==0.7.0
google-search-results==2.4.2
faiss-cpu==1.13.0
pypdf==4.0.2
emoji==2.14.0
python-dotenv==1.0.1
